First working version of web app
IT WORKS!!!!!

Currently, the web interface is able to accept a "face" image and a "test" image, which the program will attempt to pinpoint the face in. The "Submit" button will send those files to the backend, which will return the final rendered image that should hopefully show the desired face on the test image.

I have no CSS skills and barely any HTML skills, I think making this interface look any better is going to be a journey.