Create FaceRecognizerFacade main service
I didn't like that the program was split into so many modular parts and that there was no single method to orchestrate the program from start to finish (which the FileUploadController was doing). So I created a new service class for "main" methods that can simply be called from controller classes.

Also changed all facial detection and recognition method signatures to accept FaceDetectorYN and FaceDetectorSF objects directly instead of paths to the models. This ensures that not too many model objects are being created and deallocated unnecessarily, instead keeping two models around that are shared among all methods and are only deallocated when they're no longer needed.

In order to do this, I added path variables in application.properties, which are injected into the detector and predictor service classes on startup. From there, there are two simple factory methods to instantiate the detector and recognizer object, respectively.