Fix 300x300px face size restriction
The YuNet face detection model cannot discern faces that are larger than 300x300 pixels in size, really limiting the usage of the app. This was an issue especially with selfies, in which the photo resolution was high and the face in the image was well over 300x300 pixels large.

However, I made it so that the face detector performs multiple passes over every image, each time on a shrunk down version of the image. Detection results from each shrunk image are aggregated, and the result is a face detection algorithm that works with images and faces of ANY size.

In the process, I created a new MatService component, made many miscellaneous changes to almost all files, and created a more modular version of the predictFaces() method. This was all a week in the making, although I could have done all of this in 2-3 days if I didn't lose concentration so much.

Anyway, the app is much more robust than it was before and I'm happy it's the way it is.