Improve error handling and update Javadocs
This commit is basically a big clean-up of all of the backend code. I went over all of the Javadocs and updated some clearly outdated ones, as well as added some new methods that make the program easier to follow. Error handling in general is more robust now as well.