Create singleton instance of FaceRecognizerSF
I'm hoping the face recognizer is thread-safe since it shouldn't store any kind of state. So for performance, I made it into a Spring-managed singleton and leveraged Spring's dependency injection system.

I also made FaceDetectorYN into a Spring-managed prototype. As it does store an input size state, classes that require instances of it now acquire it from a factory object, similar to the factory method that existed previously.