Fix face detection algorithm
The algorithm has been partially wrong for a while; I want face detection to continue running on increasingly smaller versions of the same image until it's ran on an image < 300x300px in size, or until it's ran on an image 0.2x it's original size. I have no idea how to explain how it was behaving before, but it was just wrong. This commit corrects the loop logic and achieves the desired effect.